AB#5635306 [Ibizafication][FunctionsEditor] Change getFunctionUrl Dialog to Callout
